Fix duplicate visitor.go.go
Removes an accidentally introduced duplicate file
